Implement node heartbeat
broadcast a <getnodeconfig> request every <POLL_INTERVAL> seconds.
if a node hasn't responded to a request in <NODE_TIMEOUT> seconds it's
status is set to uninitialized.
A node that has not responded in <NODE_TIMOUT> seconds is not considered
active and is not sent to the GUI on a <getconfigs> request.